// src/types/mock-data.type.ts
export type MockData = {
  city: {
    name: string,
    latitude: string,
    longitude: string,
  }[];
  type: string[];
  userType: string[];
  facilities: string[];
  titles: string[];
  descriptions: string[];
  previewImage: string[];
  image: string[];
  users: string[];
  emails: string[];
  avatars: string[];
};